Fix TimescaleDB migration runner and metrics insertion
Three runtime issues discovered during Docker Compose verification:

1. Migration runner: Split SQL into individual statements to avoid implicit
   transaction wrapping — CREATE MATERIALIZED VIEW WITH (timescaledb.continuous)
   cannot run inside a transaction block

2. Continuous aggregate policies: Widen refresh windows to cover at least
   2 bucket widths (TimescaleDB requirement). 5min: 10→15min start offset,
   1hour: 2h→3h, 1day: 2d→3d

3. Metrics batch insert: Cast array_fill dimension param to int explicitly
   ($6::int) — pg driver passes it as text, causing type mismatch

+++ backend/src/services/metrics-store.ts
@@ -36,7 +36,7 @@ export async function insertMetrics(metrics: MetricInsert[]): Promise<void> {
     `INSERT INTO metrics (endpoint_id, container_id, container_name, metric_type, value, timestamp)
      SELECT * FROM unnest(
        $1::int[], $2::text[], $3::text[], $4::text[], $5::double precision[],
-       array_fill(NOW(), ARRAY[$6])::timestamptz[]
+       array_fill(NOW(), ARRAY[$6::int])::timestamptz[]
      )`,
     [endpointIds, containerIds, containerNames, metricTypes, values, metrics.length],
   );
